Cleanup of writer code
