+++ backends/gcp-pubsub/write.go
@@ -1,23 +1,16 @@
 package gcppubsub
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"io/ioutil"
-	"os"
 
 	"cloud.google.com/go/pubsub"
-	"github.com/golang/protobuf/jsonpb"
-	"github.com/golang/protobuf/proto"
 	"github.com/jhump/protoreflect/desc"
-	"github.com/jhump/protoreflect/dynamic"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 
 	"github.com/batchcorp/plumber/cli"
 	"github.com/batchcorp/plumber/pb"
-	"github.com/batchcorp/plumber/serializers"
+	"github.com/batchcorp/plumber/writer"
 )
 
 // Write is the entry point function for performing write operations in GCP PubSub.
@@ -26,15 +19,15 @@ import (
 // attempt to establish a connection, parse protobuf before finally attempting
 // to perform the write.
 func Write(opts *cli.Options) error {
-	if err := validateWriteOptions(opts); err != nil {
+	if err := writer.ValidateWriteOptions(opts, nil); err != nil {
 		return errors.Wrap(err, "unable to validate read options")
 	}
 
 	var mdErr error
 	var md *desc.MessageDescriptor
 
-	if opts.GCPPubSub.WriteOutputType == "protobuf" {
-		md, mdErr = pb.FindMessageDescriptor(opts.GCPPubSub.WriteProtobufDirs, opts.GCPPubSub.WriteProtobufRootMessage)
+	if opts.WriteOutputType == "protobuf" {
+		md, mdErr = pb.FindMessageDescriptor(opts.WriteProtobufDirs, opts.WriteProtobufRootMessage)
 		if mdErr != nil {
 			return errors.Wrap(mdErr, "unable to find root message descriptor")
 		}
@@ -45,7 +38,7 @@ func Write(opts *cli.Options) error {
 		return errors.Wrap(err, "unable to create client")
 	}
 
-	msg, err := generateWriteValue(md, opts)
+	msg, err := writer.GenerateWriteValue(md, opts)
 	if err != nil {
 		return errors.Wrap(err, "unable to generate write value")
 	}
@@ -77,101 +70,4 @@ func (g *GCPPubSub) Write(ctx context.Context, value []byte) error {
 	}
 
 	return nil
-}
-
-func validateWriteOptions(opts *cli.Options) error {
-	// If type is protobuf, ensure both --protobuf-dir and --protobuf-root-message
-	// are set as well
-	if opts.GCPPubSub.WriteOutputType == "protobuf" {
-		if err := cli.ValidateProtobufOptions(
-			opts.GCPPubSub.WriteProtobufDirs,
-			opts.GCPPubSub.WriteProtobufRootMessage,
-		); err != nil {
-			return fmt.Errorf("unable to validate protobuf option(s): %s", err)
-		}
-	}
-
-	// InputData and file cannot be set at the same time
-	if opts.GCPPubSub.WriteInputData != "" && opts.GCPPubSub.WriteInputFile != "" {
-		return fmt.Errorf("--input-data and --input-file cannot both be set (choose one!)")
-	}
-
-	if opts.GCPPubSub.WriteInputFile != "" {
-		if _, err := os.Stat(opts.GCPPubSub.WriteInputFile); os.IsNotExist(err) {
-			return fmt.Errorf("--input-file '%s' does not exist", opts.GCPPubSub.WriteInputFile)
-		}
-	}
-
-	return nil
-}
-
-func generateWriteValue(md *desc.MessageDescriptor, opts *cli.Options) ([]byte, error) {
-	// Do we read value or file?
-	var data []byte
-
-	if opts.GCPPubSub.WriteInputData != "" {
-		data = []byte(opts.GCPPubSub.WriteInputData)
-	} else if opts.GCPPubSub.WriteInputFile != "" {
-		var readErr error
-
-		data, readErr = ioutil.ReadFile(opts.GCPPubSub.WriteInputFile)
-		if readErr != nil {
-			return nil, fmt.Errorf("unable to read file '%s': %s", opts.GCPPubSub.WriteInputFile, readErr)
-		}
-	}
-
-	// Ensure we do not try to operate on a nil md
-	if opts.GCPPubSub.WriteOutputType == "protobuf" && md == nil {
-		return nil, errors.New("message descriptor cannot be nil when --output-type is protobuf")
-	}
-
-	// Handle AVRO
-	if opts.AvroSchemaFile != "" {
-		data, err := serializers.AvroEncode(opts.AvroSchemaFile, data)
-		if err != nil {
-			return nil, err
-		}
-
-		return data, nil
-	}
-
-	// Input: Plain Output: Plain
-	if opts.GCPPubSub.WriteInputType == "plain" && opts.GCPPubSub.WriteOutputType == "plain" {
-		return data, nil
-	}
-
-	// Input: JSONPB Output: Protobuf
-	if opts.GCPPubSub.WriteInputType == "jsonpb" && opts.GCPPubSub.WriteOutputType == "protobuf" {
-		var convertErr error
-
-		data, convertErr = convertJSONPBToProtobuf(data, dynamic.NewMessage(md))
-		if convertErr != nil {
-			return nil, errors.Wrap(convertErr, "unable to convert JSONPB to protobuf")
-		}
-
-		return data, nil
-	}
-
-	// TODO: Input: Base64 Output: Plain
-	// TODO: Input: Base64 Output: Protobuf
-	// TODO: And a few more combinations ...
-
-	return nil, errors.New("unsupported input/output combination")
-}
-
-// Convert jsonpb -> protobuf -> bytes
-func convertJSONPBToProtobuf(data []byte, m *dynamic.Message) ([]byte, error) {
-	buf := bytes.NewBuffer(data)
-
-	if err := jsonpb.Unmarshal(buf, m); err != nil {
-		return nil, errors.Wrap(err, "unable to unmarshal data into dynamic message")
-	}
-
-	// Now let's encode that into a proper protobuf message
-	pbBytes, err := proto.Marshal(m)
-	if err != nil {
-		return nil, errors.Wrap(err, "unable to marshal dynamic protobuf message to bytes")
-	}
-
-	return pbBytes, nil
 }
