+++ backends/mqtt/write.go
@@ -1,22 +1,16 @@
 package mqtt
 
 import (
-	"bytes"
 	"fmt"
-	"io/ioutil"
-	"os"
 	"strings"
 
-	"github.com/golang/protobuf/jsonpb"
-	"github.com/golang/protobuf/proto"
 	"github.com/jhump/protoreflect/desc"
-	"github.com/jhump/protoreflect/dynamic"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 
 	"github.com/batchcorp/plumber/cli"
 	"github.com/batchcorp/plumber/pb"
-	"github.com/batchcorp/plumber/serializers"
+	"github.com/batchcorp/plumber/writer"
 )
 
 // Write is the entry point function for performing write operations in MQTT.
@@ -25,15 +19,15 @@ import (
 // attempt to establish a connection, parse protobuf before finally attempting
 // to perform the write.
 func Write(opts *cli.Options) error {
-	if err := validateWriteOptions(opts); err != nil {
+	if err := writer.ValidateWriteOptions(opts, validateWriteOptions); err != nil {
 		return errors.Wrap(err, "unable to validate read options")
 	}
 
 	var mdErr error
 	var md *desc.MessageDescriptor
 
-	if opts.MQTT.WriteOutputType == "protobuf" {
-		md, mdErr = pb.FindMessageDescriptor(opts.MQTT.WriteProtobufDirs, opts.MQTT.WriteProtobufRootMessage)
+	if opts.WriteOutputType == "protobuf" {
+		md, mdErr = pb.FindMessageDescriptor(opts.WriteProtobufDirs, opts.WriteProtobufRootMessage)
 		if mdErr != nil {
 			return errors.Wrap(mdErr, "unable to find root message descriptor")
 		}
@@ -51,7 +45,7 @@ func Write(opts *cli.Options) error {
 		log:     logrus.WithField("pkg", "mqtt/write.go"),
 	}
 
-	msg, err := generateWriteValue(md, opts)
+	msg, err := writer.GenerateWriteValue(md, opts)
 	if err != nil {
 		return errors.Wrap(err, "unable to generate write value")
 	}
@@ -97,100 +91,5 @@ func validateWriteOptions(opts *cli.Options) error {
 		}
 	}
 
-	// If type is protobuf, ensure both --protobuf-dir and --protobuf-root-message
-	// are set as well
-	if opts.MQTT.WriteOutputType == "protobuf" {
-		if err := cli.ValidateProtobufOptions(
-			opts.MQTT.WriteProtobufDirs,
-			opts.MQTT.WriteProtobufRootMessage,
-		); err != nil {
-			return fmt.Errorf("unable to validate protobuf option(s): %s", err)
-		}
-	}
-
-	// InputData and file cannot be set at the same time
-	if opts.MQTT.WriteInputData != "" && opts.MQTT.WriteInputFile != "" {
-		return fmt.Errorf("--input-data and --input-file cannot both be set (choose one!)")
-	}
-
-	if opts.MQTT.WriteInputFile != "" {
-		if _, err := os.Stat(opts.MQTT.WriteInputFile); os.IsNotExist(err) {
-			return fmt.Errorf("--input-file '%s' does not exist", opts.MQTT.WriteInputFile)
-		}
-	}
-
 	return nil
-}
-
-func generateWriteValue(md *desc.MessageDescriptor, opts *cli.Options) ([]byte, error) {
-	// Do we read value or file?
-	var data []byte
-
-	if opts.MQTT.WriteInputData != "" {
-		data = []byte(opts.MQTT.WriteInputData)
-	}
-
-	if opts.MQTT.WriteInputFile != "" {
-		var readErr error
-
-		data, readErr = ioutil.ReadFile(opts.MQTT.WriteInputFile)
-		if readErr != nil {
-			return nil, fmt.Errorf("unable to read file '%s': %s", opts.MQTT.WriteInputFile, readErr)
-		}
-	}
-
-	// Ensure we do not try to operate on a nil md
-	if opts.MQTT.WriteOutputType == "protobuf" && md == nil {
-		return nil, errors.New("message descriptor cannot be nil when --output-type is protobuf")
-	}
-
-	// Handle AVRO
-	if opts.AvroSchemaFile != "" {
-		data, err := serializers.AvroEncode(opts.AvroSchemaFile, data)
-		if err != nil {
-			return nil, err
-		}
-
-		return data, nil
-	}
-
-	// Input: Plain Output: Plain
-	if opts.MQTT.WriteInputType == "plain" && opts.MQTT.WriteOutputType == "plain" {
-		return data, nil
-	}
-
-	// Input: JSONPB Output: Protobuf
-	if opts.MQTT.WriteInputType == "jsonpb" && opts.MQTT.WriteOutputType == "protobuf" {
-		var convertErr error
-
-		data, convertErr = convertJSONPBToProtobuf(data, dynamic.NewMessage(md))
-		if convertErr != nil {
-			return nil, errors.Wrap(convertErr, "unable to convert JSONPB to protobuf")
-		}
-
-		return data, nil
-	}
-
-	// TODO: Input: Base64 Output: Plain
-	// TODO: Input: Base64 Output: Protobuf
-	// TODO: And a few more combinations ...
-
-	return nil, errors.New("unsupported input/output combination")
-}
-
-// Convert jsonpb -> protobuf -> bytes
-func convertJSONPBToProtobuf(data []byte, m *dynamic.Message) ([]byte, error) {
-	buf := bytes.NewBuffer(data)
-
-	if err := jsonpb.Unmarshal(buf, m); err != nil {
-		return nil, errors.Wrap(err, "unable to unmarshal data into dynamic message")
-	}
-
-	// Now let's encode that into a proper protobuf message
-	pbBytes, err := proto.Marshal(m)
-	if err != nil {
-		return nil, errors.Wrap(err, "unable to marshal dynamic protobuf message to bytes")
-	}
-
-	return pbBytes, nil
 }
