+++ cli/kafka.go
@@ -22,13 +22,7 @@ type KafkaOptions struct {
 	ReadGroupId string
 
 	// Write
-	WriteKey                 string
-	WriteInputData           string
-	WriteInputFile           string
-	WriteInputType           string
-	WriteOutputType          string
-	WriteProtobufDirs        []string
-	WriteProtobufRootMessage string
+	WriteKey string
 }
 
 func HandleKafkaFlags(readCmd, writeCmd *kingpin.CmdClause, opts *Options) {
@@ -59,15 +53,4 @@ func addReadKafkaFlags(cmd *kingpin.CmdClause, opts *Options) {
 
 func addWriteKafkaFlags(cmd *kingpin.CmdClause, opts *Options) {
 	cmd.Flag("key", "Key to write to kafka (not required)").StringVar(&opts.Kafka.WriteKey)
-	cmd.Flag("input-data", "Data to write to kafka").StringVar(&opts.Kafka.WriteInputData)
-	cmd.Flag("input-file", "File containing input data (overrides input-data; 1 file is 1 message)").
-		ExistingFileVar(&opts.Kafka.WriteInputFile)
-	cmd.Flag("input-type", "Treat input as this type").Default("plain").
-		EnumVar(&opts.Kafka.WriteInputType, "plain", "base64", "jsonpb")
-	cmd.Flag("output-type", "Convert input to this type when writing message").
-		Default("plain").EnumVar(&opts.Kafka.WriteOutputType, "plain", "protobuf")
-	cmd.Flag("protobuf-dir", "Directory with .proto files").
-		ExistingDirsVar(&opts.Kafka.WriteProtobufDirs)
-	cmd.Flag("protobuf-root-message", "Root message in a protobuf descriptor set "+
-		"(required if protobuf-dir set)").StringVar(&opts.Kafka.WriteProtobufRootMessage)
 }
